Server/file_service/requirements.txt
fastapi
uvicorn
python-dotenv
email-validator
python-multipart